+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.14.6] - 2021-02-11
+### Changes
+- Add BrowserExt::value(). Thanks @tdryer.
+- Refactoring derive code.
+- Update cfltk.
+
 ## [0.14.5] - 2021-02-10
 ### Changes
 - Remove unnecessary field in MenuItem.